Allow the user to write correct concrete R types
Previously, we inferred serialization types for lists and tuples from
the concrete type name. So if the concrete type name is, for instance,
"std::vector", and if that is a default list type for the given language
(say C++ here), then we serialize it as a list.

While this does allow us to serialize without knowledge of the general
type, it fails when a given concrete type maps to many general types.
This is the case with the R "list" type, which maps to morloc List,
Tuple, and Record.

Previously this worked still, because we required the user enter the R
type as "tuple" and "record". This was VERY wrong, since these are not
the R concrete types at all.

Now this is fixed, and the R types may be entered in their correct
concrete forms and serialization list vs tuple types are inferred from
the general type (which should ALWAYS be known).