(148/412 fail) Backend refactor
This was a massive refactor that is now finally passing the typechecker.
The main change was to squeeze the ambiguity out of the ExprM data
structure. This led to a great tree of follow-up changes and eventually
a near complete rewrite of the backend. Hopefully these changes will
simplify the addition of new languages in the future by 1) circumventing
the need for deeper pattern matching when translating to source language
and 2) making most (all?) invalid states unrepresentable.

But, as you can see from the 148 failing cases, the work is not
finished.