(146/412 fail) Remake invert and add Index monad
The let invert function still needs a bit of work. It creates an
unnecessary binding of from (a0 = a1). Clearly something is mixed up in
its convoluted innards.

I made a dedicated monad for making indices. It replaces the heavy use
of the MorlocMonad counter. It should be all that is necessary in the
translators now. I can probably purify still more of my code with it.

My next step will be to ensure that all the pure python tests pass.
After that, I will complete the C++ implementation and test pure C++
cases and C++/Python interop.