Found eta expansion error in test interop-3e-pp
It seems that in eta expansion step here an argument is being repeated -
the correct 3rd argument is not being added, instead the 2nd argument is
added twice. I do not understand why the old test passed, since the
numbers there match neither the correct nor the new wrong results.