(93/412 fail) Fix R support
Sort of, it still needs a bit of testing. And the let inversion needs to
be fixed. Not only are indexed variables being bound to eachother
unnecessarily (e.g., `n0 = n1` in argument-form-2-py) but also the
serialzation states of these variables is being mismatched in some
cases. For example in serial-form-11-r:

``` R
m11 <- function(s0)
{
    s0 <- rmorlocinternals::mlc_deserialize(s0, '{"name":"character"
    ,"info":{"tuple":[{"list":["character"]},{"list":["character"]}]}}')
    s1 <- list(name=s0$name, info=morloc_packMap(s0$info))
    n1 <- s1
    n0 <- n1
    n2 <- mlc_id(n0)
    s2 <- list(name=n2$name, info=morloc_unpackMap(n2$info))
    s3 <- rmorlocinternals::mlc_serialize( s2
    , '{"name":"character"
    ,"info":{"tuple":[{"list":["character"]},{"list":["character"]}]}}' )
    return(s3)
}
```

Note that `s0` appears on both sides of the asignment AND it is `s0`
(serialized name) on the left. The index should be 0 on both sides,
actually, since the data is the same, but the native and serial states
are not being tracked correctly.

`s2` also, though certainly native, is named as serial.

I should fix this and comb carefully through all generated code before
continuing.

Or rather, I should get C++ working. It is almost there. And it gives
sometimes useful type annotation info. Keeps me honest.