(42/412 fail) Fix recmap collection
Also I added a defaultValue instance for FoldManifoldM. This instance
works only when all 6 types have the same acumulator type and when this
type is a Monoid. Though specific, this case is very common. I use it,
for example, in any sort of query that collects lists, maps, or
sets of values from across the structure.