(16/412 fail) Stop needless context passed in lam
So the problem in eta-reduce-8 is that the args are not being converted
to the right serialization state. For instance, look at this generated
python code:

```
def m1(s0, s1):
    ...
    n12 = onThree(lambda s2, n3, s4: m2(n0, s2, n3, s4), n11)
    ...

def m2(n0, s2, n3, s4):
    n7 = m9(n0)
    n8 = (n7, n3)
    return(n8)
```

See that `s2` and `s4` are passed from the lambda. This cannot be
correct. Lambda terms are *always* native. This test passes in Python
because Python does care that s2 and s4 are the wrong type when passed
to m2. They are not used. But in C++, the wrong type is generated for
these (they are both strings) so the code fails at compile time.