(151/412 fail) Stricter ManifoldForm types
ManifoldForm now takes two parameters, one for contextual arguments that
are applied to the manifold and one for bound arguments that are
supplied by the function through a lambda.

More importantly, the bound arguments can be passed as both serial and
native, this will avoid having to reserialize at each usage.

I also added a specialized bifunctor class (with added monadic cases)
and a bifoldable class. These simplify working with ManifoldForms.

Obviously, the kinks are not all worked out, but at least it passes the
typechecker now.