(30/412 fail) Nearly all non-interop tests pass
There are two non-interop tests that fail: edge-cases-2 and
eta-reduction-8-cpp. The latter is related somewhat to `specialize`
where the `Nothing` mapping to a passthrough fixes the Python problem
(but the types are still not working out for C++, so something is off).

The interop tests are failing (probably) just because I don't yet
deserialized foreign calls. I probably just need to add a naturalizeN
call to every foreign call. That is not ideal for all cases, since
sometimes deserialization may not be necessary.