(15/412 fail) All non-interop tests pass
This is actually a huge step forward even though 2 more tests fail
overall. The hard tests that initially inspired this long refactor now
pass. The interop tests that are failing now passed in the recent past.
The problems with them seem to be trivial mis-wiring. Or at least, that
is my sincere hope.

In interop-3a-pr, it appears that an R type is leaking into Python. This
leads to a serialization error where no packer for type "numeric" can be
found.