(7/412 fail) Fixed express typing problems
Also added safeZip all through generate (I should probably do this
everywhere zip is used). I replace the old "equalZip*" family with
"safeZip" that returns Maybe instead of raising an error. If necessary,
I can always use `fromJust` at the call site. Doing this ensures the
link in the ghc error message points to the right line of code (rather
than the line of code in the Internal library).

The remaining test seem to fail when a manifold inherits both serialized
and unserialized forms for the same argument id. The functions are
trying to deserialize these.