All tests pass
The last failing test was due to an indexing problem. When I selected a
maximum index in `Common:maxIndex` that would be used for generating the
let terms, I did not include indices that appear in manifolds arguments
but not in bodies. This case appears when an argument is passed into a
composition but never used. Naming conflicts cause problems in some
languages but not others.

What's next? I need to test my case study. If that passes, then morloc
is good to go, for now.