All tests pass
The test `interop-10` failed because the typechecker was not
distinguishing between variable terms in different languages. When a
variable was encountered in the synth step, the variable was looked up
by name in Gamma. The variable was unified with the first matching
element in Gamma. However, in certain cases, the first matching element
was not in the same language. In the `interop-10` case:

```
foo x y = f g (h (x, y))
```

Where f and g are C++ functions and h is a Python function.

The existential variables x and y are created as C++ vars in the
top-level lambda binding. Then they are synthesized in the python
environment, but when looked up are assigned to C++ existentials. If
instead, the expression where `f g (h x y)`, then they would be
*checked* against python existentials generated for the type of h.

The reason that the expression `f g (h [x])` passed, is that the
synthesis case of lists generates a new existential and checks it
against elements of the list, this avoids the C++ association.

I solved the problem with `interop-10` by making the lookup
language-specific. Now, in synth VarS a new existential variable will be
generated for each language instance of x (C++ and Python, in this
case).

However, this change exposed another bug causing the tests
eta-reduction-1 and eta-reduction-2 to fail. `eta-reduction-2` has the
form:

```
bar x y z = add x . add y . add z
```

Where each add has implementions in all supported languages. In the
realization step, a language was selected for each expression in the
composition. The algorithm I was using assigned base expressions
penalties of 0 and added penalties upwards as functions were called.
However, when I make selections downwards, these base terms where all
being assigned to python types since python happened to be the first in
the list and all scores where equal (all were 0). I fixed this with a
kludge: adding a penalty of 1 when languages of the child element
differed from the parent.

Previously, this second bug was hidden by the first bug. Looking up the
first variable, regardless of language, happened to find the variable in
the right language.

I should probably work through this more systematically, but I am short
on time (as always).