(1/417 fail) Add test for realization of packers
Currently concrete type inference does not extend to packer functions.
This means that code like the following fails:

```
type Py (Map a b) = "dict" a b

unpackMap :: unpack => [(a,b)] -> Map a b
packMap :: pack => Map a b -> [(a,b)]
```

Instead we need to write a signature for each language.