Add test and code for concrete packer inference
I do concrete type inference in Desugar for (un)packers.

All is not well, though, since my changes broke the flu case study in a
new way. Now the serialization engine is confusing the RootedTree and
UnrootedTree packers. These both share the same concrete type for R.

So next I need to write a minimal test for this and then fix the bug. I
suspect something is off in my new packer code in Desugar.