(356/407 fail) Refactor type defs
Now types are defined as follows:

```
type (Point a b) = (a, b)
type Py => Int = "int"
```

This is just a difference in syntax, but the idea is that the language
is a namespace for the type functions. We create language-specific types
by evaluating the type in the scope of the language type function
namespace. And these may be seen as functions with the generics serving
as arguments.

To facilitate downstream conversions, I load concrete and general
typedif maps into the MorlocState record in the MorlocMonad. These have
the type:

```
data MorlocState = Morloc State
  { ...
  , stateConcreteTypedefs :: GMap Int MVar (Map Lang (Map TVar [([TVar], TypeU)]))
  , stateGeneralTypedefs  :: GMap Int MVar           (Map TVar [([TVar], TypeU)])
  , ...
  }
```

The `Map TVar [([TVar], TypeU)]` type is a scope. It can be used to
either simplify a general type (resolving all type aliases) or to
generate a language-specific type.

Next I need to figure out how to handle (un)packing. I am inclined to
create a map of signature properties:

```
data MorlocState = Morloc State
  { ...
  , stateInnerMogrifiers :: GMap Int MVar (Map Property [(EVar, TypeU)])
  , ... }
```

Of the properties, `pack` and `unpack` are just two. More may be added.
Evaluating a type in the scope of inner mogrifiers will yield not just a
new type but also a composition of functions to enact the change. The
principle is the same as what is currently being done for generating
(de)serialization code. Generalizing the approach will simplify
(de)serialization and finally allow powerful features I've considered
since the beginning, such as automatic test generation, type casting,
caching, UI generation, and other applications of automatic code
generation.