(125/400 fail) All systems defined
There are multiple logical issues to work out. A major change for the
user is that now all types that are used in the program, even those that
are somewhat invisible in the serialization code, must be imported. This
is a bit tedious. It is, however, I move in the right direction. It
allows plains to be designed with different type conventions. The
compiler should assume as little as possible about a language.