(121/400 fail) Special (un)packer key extraction
One of the goals of this refactor is to remove the special status of the
(un)packers. They are just one of several inner transmogrifiers (though
currenty the only that are used). However, the inheritance patterns are
a bit complex. A packer has forms like `List (Tuple a b) -> Map a b` and an
unpacker has forms like `Map a b -> List (Tuple a b)`. Both are imported if
`Map` is imported. That is, they are not explicitly imported by function
name. But determining whether to import `Map` or `List` requires knowing
whether to derive the key from the input type or output type. So it is
dependent on the property Pack/Unpack. And what of other
transmogrifiers? Will they also have their own special handing?

 Explicit is better than implicit
 - Zen of Python

 - pack - [2] make a datatype from simpler components
 - unpack - [1] simplify a datatype for interop
 - cast - [1,2] convert between types
 - sample - [1] generate data from types

All of these are type functions. They all have instances for many types.
So type classes may be relevant. In Haskell, you could do this:

```
{-# LANGUAGE MultiParamTypeClasses #-}
{-# LANGUAGE FlexibleInstances #-}

import Data.Map (Map, fromList, toList)

class Packable a b where
  pack :: b -> a
  unpack :: a -> b

instance (Ord a) => Packable (Map a b) [(a, b)] where
  pack = fromList
  unpack = toList

instance Serializable a b where
  serialize :: a -> b
  deserialize :: b -> a

instance Serializable Int String where
  serialize = show
  deserialize = read

instance (Serializable a) => Serializable [a] String where
  serialize xs = list (map serialize xs)
  deserialize x = map deserialize (split "," (unbracket x))

instance Serializable Int Bytestring where
    ...
```

The transmogrifiers, though, are not explicitly written by the user, but
rather used by the compiler to accomplish certain tasks, such as
serialization, caching or test generation. So maybe they are special
type classes? Or will be, once I add type classes.

So, adapting the Haskell code from above, the morloc compiler would use
`Packable` and `Serializable` to generate serialization code. This seems
pretty doable. So my current transmogrifier/property scheme is just a
special case of typeclasses.

The serialization scheme requires a parameter for the serialization
form. This could a string, as is done when serializing to default
JSON. It could alternatively be some other binary format. The typeclass
approach allows us to specify in code, rather than in the compiler, how
serialization is handled.

But not yet. I will continue with the limited transmogrifier concept, work out
the bugs, and then generalize to full type classes.