(116/400 fail) Fail if concrete type cannot be inferred
Some tests fail now that previously passed. This is true for dynamic
languages where I was previously giving them the wrong type.

For example:

```
import pybase (map, add)

type Py => List a = "list" a

foo = map (add 2.0) [1,2,3]
```

Here when we try to generate the concrete type for `foo`, we do the
following:

List Real -> "list" Real

Using the type functions `List a = "list" a`. But then we find no rule
for substituting `Real`, so we are left with a type that is half
concrete and half general. However, in the python code, the type is
never used so the build still succeeds.

My solution was to tag all type functions as terminal or non-terminal.
Terminal ones are concrete functions such as

```
type Py => Int = "int"
type Py => List a = "list" a
```

Whereas non-terminal ones are:

```
type Py => Map a b = [(a, b)]
type Int64 = Int
```

The distinction is only needed when mapping between namespaces, such as
from one language to another or from general to concrete. Within
namespace transformations will be handled gracefully in any case.