(51/400 fail) Factored Desugar into TypeEval.hs and Restructure.hs
The "Desugar" module was not doing any desugaring. In the past it fully
evaluated all the types. What I then considered "desugaring" since I
thought of the type functions as just aliases. Now this evaluation is
postponed. But all the evaluation code remained the "desugar" module.
The main purpose of the module was to move info in the expression tree
into state. So I split the module in two.

Morloc.TypeEval contains functions for evaluating types within a given
scope (or two).

Morloc.Frontend.Restructure exports a single functions, restructure,
that moves packers, mogrifiers, sources, type signatures, and more into
state. It is used just once in the frontend after the parser.