(40/400 fail) Fix bugs from f8beb7 and 95ee67
These two tests, packer-definitions-3 and packer-definitions-4 were
added on 2023/12/01 and 2023/12/03 respectively.

These were both new tests that checked for bugs that appeared in the
case study (if I recall correctly).

I did not document properly what specific problem gave rise to the
tests. Bad form. In the future, I should be more disciplined about
submitting issue reports on GitHub for each new test I add.

The code I changed is in the `express` match against `LamS`. Here I want
to replace the body index with the lambda index and remove the lambda.
This may be done safely since the argument the lambda passes in has been
saved in the argument list associated with the body expression (in
`parameterize`). The bug in the code occured because I moved over not
only the lambda index but also the lambda type.