(13/400 fail) Pushed the indices deeper
In the past, I managed serialization within the scope of the manifold.
But this has never been the right approach. The manifold may contain
expressions from many scopes. For serialization, this wasn't normally a
problem. I usually had just one (un)packer pair for each data type, so I
would pool them globally. But I am beginning to do more type
manipulations that require calculations across scopes. The link to these
original scopes are stored in the indices I couple to expressions in the
parser. These indices were being lost when I convert from SAnno to
PolyExpr. Now I propagate them all the way down to MonoExpr. But I need
to go further still. They need to go all the way to the end. Then I will
be able to (de)serialize any expression since I will have a handle on
its scope. It does require a bit more book-keeping. And I need some more
formal notion of scope transfer. When I create new expressions, I need
to generate new indices and determine how to derive the new index from
an older one.

But first, I need to fix the record handling. These are the last tests
that are failing. They fail because subtyping is breaking when it, for
example, compares "Person a" to "Person {info::a, name::Str}".  It sees
"Person a", correctly, as a parameterized type and "Person {...}" as a
NamType. These are not directly comparable. What needs to be done is the
partial evaluation of "Person a" into a record using its type
definition. But when should these terms be evaluated? And how far?
Previously I evaluated everything all the way down. But there is value
in not doing so. Then I can map "Person a" to multiple things.

I can extend the subtype function to attempt partial evaluation of types
when needed within the local scope. This requires passing scope into the
function, and it requires deciding when to partially evaluate. Thinking
through these problems will be the topic of the next commit (or two).