All tests pass, but at what cost?
There is a flaw in my "reasoning" about type scoping. Just evaluating
the type in the scope where it was defined is insufficient. It fails for
generic types that are inferred in higher scopes.

For now, all tests pass because I've made two bad compromises:

 1. I evaluate all types in a global scope, where all type definitions
    defined in any module that is used by the program are in scope. If
    any type definitions contradict eachother, an error is raised.

 2. I fully resolve all general types before typechecking. I toyed with
    the idea of not doing so, and resolving types as needed. But this
    ended up not working because of the scoping issues. There isn't a
    fundamental reason why delaying is impossible, I just need to
    figure out the math. Delayed evaluation allows type functions
    intermediate types. This is not something I've ever used, but it
    seems as though it might have appliations ... somewhere.

```
module c (f)

f :: a -> b -> C

module b (h)

h :: B -> B

module a (foo)
import c (f)
import b (h)

foo x = f (h x) 42
```

The type of `f` is inferred in module `a`, defined in `c`, and depends
on `b`. We should not have to import type B from b since we do not
directly use it. We should also not have to import C from c. The
requirement to import all these would make for very bloated and ugly
code. Instead, the types should be looked up when needed.

I eventually need to return to this and really formalize the behavior.
But for now with the two compromises everything works.