Add test and handling for specialized (un)packers
Take for example the Map type. Most generally, it will have two
generic parameters, a key and a value type: (Map key val).

The packer type will be:

```
packMap :: pack => ([key],[val]) -> Map key val
```

But suppose we want to define a packer for the special case when the key
type is a string. Indeed, for some languages, the string key may be the
only supported type. Such languages support only the `Map Str val`
subset of types in `Map key val`. This should be legal.

I've now fixed the code so that this is permitted and added a test,
`packer-definitions-5` that covers it.