Fixed bug in Python and R serializers
Trivial bug. There was a guarded function that halted
recursion under some condition and then the main recursion function. The
two need to alternate so the condition is always checked. But in one
case I recursed directly into the unguarded funtion thus skipping the
check. This lead to an error case.