Allow nexus inputs to be files
This avoids the extra level of quotation required in bash as well as
being something that is required for practical use.

I am currently slurping the entire files into memory and then passing
them as strings to the pools. This is not great, but until I have
streaming working, it is the way it has to be. Of course, I could
alternatively pass filenames in as strings.