Add a borked Dockerfile
It mostly works, honestly, you'll just have to run half of it or so and
then do the rest manually. There is something weird that breaks when I
try to install stack. It works when I run the code in a bash instance
inside the container, but it breaks when the same code is run by the
build process. Meh, don't know don't care, this is a "book" problem
(what do I mean by that? Just read the blog post I will write at some
point in the future).