Begin dev cycle 0.44
In this dev cycle I want to add very basic support for typeclasses. I
will allow type classes to be defined, which will offer parametric
polymorphism, but I will not add a constraint system.

The most pressing use case of this is for serialization. Currently I do
this:

```
packMap :: pack => [(a,b)] -> Map a b
```

Here the `pack` is a special term the states, in essense, that this
function is specializing some generic function `pack :: a -> b`. But
this is a rather odd syntax. A function can only specialize one
function, so presumably I can't use multiple such constraints, and
should I raise an error if someone tries? Worse, I can't cleanly couple
multiple functions that go together in a typeclass. In this case, the
`pack` and the `unpack` instances are separate and I have to unify them
with (rather tedious) outside code later.

With typeclass syntax I can do:

```
class Packable a b where
    pack :: a -> b
    unpack :: b -> a

instance Packable [(a,b)] (Map a b) where
    source Cpp from "map.hpp" ("pack", "unpack")
    source Py from "map.py" ("pack", "unpack")
```

There are a few strong limitations I am going to place on the typeclass
system for now.

First, only source statements are allowed in instances. The functions
written in typeclasses will tend to be very basal functions.  In
Haskell, additional functions are often defined inside the typeclass.
This is useful, but not necessary. I can define these functions outside.

Second, typeclasses may not be used as constraints. This is simply
because I am lazy. Adding typeclass constraints to signatures does
improve readability, at times, but it requires additional checking of
things that can be inferred. I will add it later. In fact, I already
support it syntactically, so knock yourself out.

Third, typeclasses and instances are global. This is purely laziness on
my part. But my scoping system is currently borked and needs to a total
rewrite, and I don't have the time. Transitioning to scoped in the
future will require a bit of finagling with the import/export statements
-- nothing radical. Further, it is probably best practices to not have
different typeclass instances existing in the same program anyways.
Making typeclasses and instances global also is a solution to the orphan
instance problem (though an admittedly an ugly one).

```
class Arithemtic a where
    add :: a -> a -> a
    sub :: a -> a -> a
    mul :: a -> a -> a

class Zero a where
    zero :: a

instance Zero Int where
    ...

instance Zero Real where
    ...

instance Arithmetic Int where
    ...

instance Arithmetic Real where
    ...

foo :: [Int] -> [Int]
foo = foldr add zero
```

When the typechecker encounters a symbol, it will look up a type
signature for it. If it is term defined in a typeclass, the returned
type will be generic (usually, or at least it will be whatever was
defined in the `class` block). The types must be inferrable from the
context without considering the instance data. Technically, it will
sometimes be possible to infer the instance by ruling all but one out,
but I will not yet do this.

The instances will be used when need to look up sources. Then we will
search by the function name and the value of the class parameters. For
example, for `pack` we look up the types `[(a,b)]` and `Map a b` among
the instances `Packable`. Now, these will be fully inferred, so we will
need to search by subtyping.

I more or less already have a typeclass system for the special case of
(un)packers.

I may also need to create a Prelude with typeclass definitions to
provide some rigidity to the system. At very least, the Packable
typeclass needs a definition as do future internal typeclasses.