(1/408 fail) Add typeclass info to state
Next I need to look up typeclass functions in the frontend checker
at the point where I resolve terms. Then I can update `stateSources`
with links to the instance sources for each term.

That should be all that is required. In the generator all that matters
is the inferred general type (which is stored in the SAnno object) and
the index to the source.

A little additional work will be needed if I want to validate the
correctness of any given or omitted typeclass constraints.