All tests pass
The prior commit actually was correct, it was my test that was wrong.
Well, that and uninformative error messages. Handing incorrect code is
the vast hole in my language -- there are so many more ways to be wrong
than right -- for now I just blame it all on the user. Maybe that is way
I don't have any users? No, that isn't the reason. No one has ventured
far enough into the morloc labyrinth to step on the glass I've strewn on
the ground. Except for me, of course, I've scampered all over the glass.

The typeclass handling code is actually quite intuitive and beautiful.
All instances are checked, so there should be no room for lurking
invalid but unused code, and nothing is checked more than once. All the
checking is elegant with no hacky bits. I like it. And typeclasses are
fully considered and evaluated away in the frontend typechecker, so no
new complexity is added to the backend.

Next up, remove all the special handling for the packers.