Separate typeclass and merge logic from Treeify
Treeify turns Expr's created by the parser into the SAnno/SExpr tree
used in the typechecker. Previously, it also collected all the
typeclasses, now I am breaking that logic into a new module. But the
new Typeclass module and Treeify use some of the same merging logic, so
I also created a new Merge module. The merge logic is very important in
morloc, since there are many implementations allowed, so it is
reasonable to have a dedicated module where the handling can be
carefully developed and unified.