v0.45.0 - explicit instance imports
In Haskell, all instances of polymorphic terms are imported implicitly
when the defining module is imported.

In Morloc, however, since terms are polymorphic both in type and in
language, it is often useful to more selectively import instances. So
now any terms defined in typeclasses must be specified. For example:

```
import base (show)
```

Where `show` is method of the `Show` typeclass.

Another part of my rational is that monomorphic and polymorphic
functions ought to be treated as similarly as possible. That said, even
finer control might be worthwhile. We could require that typeclasses are
imported explicitly, forexample:

```
import cppbase (Ord)
```

Or to import only specific terms:

```
import cppbase (Ord(ge, lt))
```

This would also allow one to make aliases for the polymorphic terms.