Rename "plain" to "plane"
Damn synonyms refer to wide flat things. I guess the morloc levels are
planes of the mathematical or astral kind not plains of grass. I rather
like plains of prairy grass, most of all in the weeks of spring when
they wave softly and greenly in the wind and smell of honey. Morloc
planes are black and white and smell of nothing.