Start dev cycle 0.47.0
Consider this code:

```
import base ("fold", "add")
source Py ("sum")
sum = fold add 0
```

Here I define a specialized Python-specific summing function as well as
a general function of the more basic fold and add functions. In this
case, the python source `sum` may be used if the context favors it, or
in other language contexts the general `sum` composition may be used.

This is good. It is one of the motivating features of `morloc`.

But the system is currently too permissive. The following is allowed:

```
module m (x)
x = 1
x = 2
```

Both implementations of `x` are accepted and the latter is arbitrarily
selected. This should raise an error because the values conflict.

The same problem may occur in more deeply nested cases:

```
x = add 0 1
x = add 0 2
```

```
x = 1
x = add 0 1
```

```
x = add 0 1
x = add 0 (add 0 1)
```

I can resolve this family of errors with a value checker that checks
pairs of implementations for data conflicts.

Other issue is when multiple implementations exist for a given language
with no rule to distinuish them. For example:

```
source Py from "foo.py" ("bar" as f, "baz" as f)
```

Or here is a similar example:

```
x = foldr add 0
x = foldl add 0
```

Here there are two source definitions of a Python function. Even after
language resolution, there are two defintions. Further, the compiler
knows nothing about these functions. It has no particular way to choose
one of them. Currently the last one defined is used. But this makes the
first definition unreachable. Should unreachable code even be included?
In the future, we could attempt compilations and if they fail reach for
alternative functions. Or we could benchmark each function an choose the
faster one. Or we could statically analyze each and choose the better
one. Or we could consider the dependencies each requires. There are many
possible future avenues for research. But at the moment I do none of
this. Instead, I should send the instances to a rule-engine where future
research may be conducted but that, for now, just fails with the
conclusion that no rule exists to distinguish the functions.