Implement basic value checker
All prior tests pass. This only means that the value checker isn't
killing any of the correct code I wrote in old test. This is not
surprising.

I need to write new tests to specifically address the kinds of errors
that the valuechecker catches. Such as

```
x = 1
x = 2
```

These should be unit tests over the checkPair function. I should ensure
that I have a test for every location in the code that raises an error.