Make lite Dockerfile for compiler and lang support
I've cut the weight from 6.9Gb to 1.3G. Most of this comes from using a
two-stage build that eliminates all the Haskell build artifacts.

The container also includes all support for C++, Python and R and the
core modules needed for running all tests.

I also threw in vim and some setup scripts so it is a little more
hospitable to visitors.