Solve the lgmp not found problem
When I converted the build to static, as was necessary for making the
small docker images in the two-stage build, I could no longer compile on
arch. It ends up that the problem was that the static build required a
static `libgmp.a` libraries and I only had `libgmp.so` shared libraries.

Making this work on arch required installing libgmp-static from AUR.