(5/427 fail) Propagate qualifiers to code gen
There is one `check` step in the typechecker that removes universal
qualifiers. At that point, I can store the removed qualifier in a list
in state. Then after typechecking, I can apply the learned type context
to the qualifier existential variable to get the inferred type.

This inferred type I then retrieve from state by general index in the
Generate.hs module when I create `AppSrcN`. This is the one place where
the C++ code generator will need the qualifier instance. It is the
constructor that stores the inforation needed to make the name of a C++
function. Now it will generate C++ functions like `snd<int,bool>`,
rather than just `snd`.

5 cases still fail. The reason they fail is that the order of morloc
universal qualifiers is dependent on the order of evaluation of the type
signatures in the parser. If the order of qualifiers does not match the
order of template arguments in the C++ source code, then the incorrect
template arguments will be generated. I can remedy this by allowing
`forall` syntax for introducing qualifiers.

So rather than write

```
snd :: (a, b) -> b
```

I would write

```
snd :: forall a b . (a, b) -> b
```

This is wordy, but it orders the qualifiers. It makes it clear that we
are interested in:

```
template <class T1, class T2>
T2 snd(tuple<T1,T2>)
```

But not:

```
template <class T1, class T2>
T1 snd(tuple<T2,T1>)
```