(4/427 fail) C++ template args must match morloc generics
Morloc will generate template arguments if the signature for the
function is generic. The string-encoding test was failing under this new
implementation because the morloc function was generic (a -> a) but the
C++ function was not (string -> string). Then the morloc compiler generated
the function with a <string> template argument. The C++ compiler then
died because it knew that this function accepted no such argument.

I think this higher level of strictness is a good thing. The relevant morloc
priniciple is that the source code functions are designed to match the
morloc signatures. So raising errors when the disagree is the right
choice.