(1/427 fail) `forall` syntax for setting order
Suppose we have this function:

```
bar :: b -> (a, b) -> b
```

And this implementation:

``` C++
template <class A, class B>
B bar(double dummy, const std::tuple<A,B> &x){
    return(std::get<1>(x));
}
```

The morloc parser will identify the generics in `bar` in the order they
are encountered. So it will find the terms [b,a] and generate the type:

```
forall b a . b -> (a, b) -> b
```

Then the [b,a] parameters will be inferred from context to, say,
[double,str]. Then when this `bar` is used in morloc code, it will be
written as `bar<double,str>`. However, this is backwards since C++ wants
[A,B] order.

To deal with cases like this where order of qualifiers matters, I've
added parser support for the `forall` term. So now we can write `bar`
as:

```
bar :: forall a b . b -> (a, b) -> b
```

There is more that can (but maybe shouldn't) be done with this syntax.
Generic terms that are not used in the type may be introduced. For
example:

```
bar' :: forall a b c -> b -> (a, b) -> b
```

But this not a haunted opera house that I feel like venturing into just
yet.