Replace forall with qualifier argument syntax
Originall I would handle generics implicitly, for example:

```
map :: (a -> b) -> [a] -> [b]
```

This is simple, however, the mapping to positional template arguments in
C++ is ambiguous.

To force a specific order, I added a `forall` syntax:

```
map :: forall a b . (a -> b) -> [a] -> [b]
```

While the meaning is clear, it is a bit wordy (adding both `forall` and
a new use of `.`). Writing this for every function is a bit tedious.
Writing it only when necessary, gets arbitrary and confusing.

So here I am converting to a third options: type argument syntax.

```
map a b :: (a -> b) -> [a] -> [b]
```

With this syntax, the role of `map a b :: ...` as a type function is
clear. It parallels the definition of the `map` term function. Further,
it provides a clean syntax to talk about the function type. We can say
that `map a b` is resolved to, for example, `map Str Int`.

The syntax also suggests ideas for type functions. For example:

```
mapS b :: map Str b
```

The function mapS is the family of functions that map a list of strings
to something. This syntax is partially redundant, since we already have
type definitions, such as:

```
type MapF a b = (a -> b) -> [a] -> [b]
map :: MapF a b

type MapSF b = MapF Str b
mapS :: MapSF b
```