(All fail) Partial socket support for Python and C
I converted a simple morloc program with Python and C interop over to
socket-based IPC. Now I am working on porting the code back into Morloc
source.

Here is how the program works.

 1. The user compiles the morloc script (same as before)
 2. The user interacts with the nexus.py script (same as before)
 3. When called with a subcommand and arguments, the nexus.py script:
    1. Initializes the pool for each language as a background process.
    2. Creates a domain socket connection to the pool that implements
       the top function used by the subcommand
    3. Sends a command to the pool with the manifold ID and an argument
       list. Where every argument is a "key" (currently a filename for a
       temp file). The keys can be used to access the serialized data.
    4. Listen to the socket until a response returns (which will be a
       single key to the output data)
    5. Open the key and print the result to the user

So the UI is the same as before. Now, a little more detail on the pools.

Each pool is a domain socket server. It listens asynchronously for
commands from any client. The client may be the nexus or another pool.
Either way, the command format is the same: A space-delimited list of a
manifold ID followed by argument keys. The messages sent over the
sockets are always short and fit into a 1024 byte buffer. Note that
creation of keys is handled by morloc. If keys are 20 characters, then
the buffer will handle functions with 50 arguments, far more than is
sane. Still, I should add checks on this. Maybe within the compiler I
can raise an error if more than, say, 20 arguments are given to a
function.

Anyway, when a client sends a command through the socket to the pool,
the pool asynchronously starts a new job in a new process. This should
support parallelism, though I haven't tested it yet. When the job
finishes, the pool sends the key to the data back to the client over the
socket. Asynchronous systems like this can have subtle bugs. I'll need
to really heavily test this.

The socket communication and language servers eliminate the need to
startup many interpreters (costly in the cases of R and Python). It also
offers about the fastest transmission and lowest latency possible
without shared memory. Domain sockets are faster than TPC, for instance.

There are two more components of this architecture that I need to add.

First, I am still writing data to files on the disk. The "keys" I pass
around are filenames. This is deadfully inefficient and insecure. I can
replace this approach by using a database. My current idea is to use
RocksDB to store all intermediate serialized data. RocksDB is a
key-value database. I can make a random binary key and associate it with
a random binary blob of data. And that's that. Nice and simple. RocksDB
is an embedded database. So the nexus would initalize an instance and
the pools would communicate with the nexus over domain sockets to put
and get data. RocksDB caches data in memory when possible but writes to
disk when needed. We can set the size of the memory cache.

Second, I am currently still using rather hacky and slow methods for
serializing to JSON. I want to replace this with a better serialization
solution. Probably MessagePack. This will not be a massive change
internally. Most of the complexity of morloc serialization is in
wrangling the data down to the primitive form that has a one-to-one
mapping to JSON. This will not change. Just the final step, where I
currently pass the JSON-compatible data forms to some language-specific
library that I wrote, I'll pass it to a MessagePack library. This will
not be as simple as a calling a "serialize" standard function on my
data. I'll need to customize the serialization to ensure that all
languages are handling all data types in the same way. So the
language-specific modules I currently use may not disappear completely.

Anyway, all that is for later. Right now, I need to wrap up the socket
programming. There is a fair bit of wiring left to do and I haven't even
started with R.